Patch use of HexHex in cards
