+++ CHANGES.txt
@@ -3,6 +3,7 @@ protograf Changes
 
 2026-00-00 (0.5.4)
 ------------------
+- Patch use of HexHex on cards
 - [PLANNED-1] Foldable card box
 - [PLANNED-1] Structure: break-up proto into smaller modules
 - [PLANNED-2] Docs: list all props for Hex, Circle, Rectangle, Line
